Factoria de pools creado
